+++ core/CHANGELOG.md
@@ -6,9 +6,14 @@
 
 - Move `Executor` to `libp2p-swarm`. See [PR 3097].
 
+- Update `rust-version` to reflect the actual MSRV: 1.60.0. See [PR 3090].
+
+- Update `multistream-select` to `v0.12.1`. See [PR 3090].
+
 [PR 3031]: https://github.com/libp2p/rust-libp2p/pull/3031
 [PR 3058]: https://github.com/libp2p/rust-libp2p/pull/3058
 [PR 3097]: https://github.com/libp2p/rust-libp2p/pull/3097
+[PR 3090]: https://github.com/libp2p/rust-libp2p/pull/3090
 
 # 0.37.0